# openapi.yaml
swagger: '2.0'
info:
  title: Bot-API-Gateway
  version: 1.2.0
host: bots-gateway-prod-83hfw8tp.ew.gateway.dev
schemes:
  - https
...
paths:
...
  '/activate/{key}':
    get:
      summary: Activate device
      operationId: activateDevice
      parameters:
        - in: path
          name: key
          required: true
          type: string
      x-google-backend:
        address: https://bot-service-634498283437.europe-southwest1.run.app/
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        '200':
          description: OK
...